requirements/base.txt
Django==5.0.7
djangorestframework==3.15.2
django-cors-headers==4.4.0

# Banco de dados
psycopg2-binary==2.9.9

# Cache e filas
redis==5.0.8
celery==5.4.0
django-celery-beat==2.7.0
django-celery-results==2.5.1

# Configuração
python-decouple==3.8

# Arquivos e imagens
Pillow==10.4.0
django-storages==1.14.4

# Segurança e auditoria
django-simple-history==3.7.0

# Utilitários
whitenoise==6.7.0
django-filter==24.3

# PDF
reportlab==4.2.2
weasyprint==62.3

# Formulários
django-widget-tweaks==1.5.0